move the code to the actual entrypoint for node
