+++ node-src/tasks/index.ts
@@ -11,7 +11,6 @@ import snapshot from './snapshot';
 import storybookInfo from './storybookInfo';
 import upload from './upload';
 import verify from './verify';
-import { getBranch, getCommit, getSlug } from '../git/git';
 
 export const runUploadBuild = [
   auth,
@@ -31,20 +30,3 @@ export default (options: Context['options']): Listr.ListrTask<Context>[] => {
 
   return options.junitReport ? tasks.concat(report) : tasks;
 };
-
-export type GitInfo = {
-  branch: string;
-  commit: string;
-  slug: string;
-};
-
-export async function getGitInfo(): Promise<GitInfo> {
-  const branch = await getBranch();
-  const { commit } = await getCommit();
-  const slug = await getSlug();
-
-  const [ownerName, repoName, ...rest] = slug ? slug.split('/') : [];
-  const isValidSlug = !!ownerName && !!repoName && !rest.length;
-
-  return { branch, commit, slug: isValidSlug ? slug : '' };
-}
